LayoutTests/imported/w3c/web-platform-tests/screen-wake-lock/wakelock-type.https.window-expected.txt


PASS 'type' parameter in WakeLock.request() defaults to 'screen'
PASS 'TypeError' is thrown when set an invalid wake lock type
